Split controllers, models, and routes into separate files
