+++ server/server.js
@@ -5,18 +5,9 @@ const morgan = require('morgan');
 // eslint-disable-next-line no-unused-vars
 const colors = require('colors');
 const express = require('express');
-const pg = require('pg');
-const bcrypt = require('bcryptjs');
-const { v4: uuidv4 } = require('uuid');
+const usersRouter = require('./routes/users');
 
 const app = express();
-
-const connectionUrl = 'postgres://rdaasabl:U4aOtl4SwgJbehXGxGdM-5XLYvR9EU4r@ruby.db.elephantsql.com:5432/rdaasabl';
-
-// Connect to our database
-const pool = new pg.Pool({
-  connectionString: connectionUrl,
-});
 
 // Body Parser
 app.use(express.json());
@@ -26,65 +17,8 @@ if (process.env.NODE_ENV === 'development') {
   app.use(morgan('dev'));
 }
 
-// Middleware to check validity of username
-const checkUsername = (req, res, next) => {
-  const user = req.body.username.toLowerCase();
-  const text = `SELECT * FROM ubiquitous_spoon.users WHERE username = '${user}'`;
-  pool.query(text, (err, response) => {
-    if (err) {
-      return next(err);
-    }
-    if (response.rows.length === 0) {
-      res.locals.username = user;
-    } else {
-      return next({ log: 'checkUsername', message: { err: 'Username already exists' } });
-    }
-    next();
-  });
-};
-
-const checkPassword = (req, res, next) => {
-  const { password1, password2 } = req.body;
-
-  if (password1 !== password2) {
-    return next({ log: 'checkPassword', message: { err: 'Passwords must match ' } });
-  }
-
-  bcrypt.genSalt(10, function (err, salt) {
-    bcrypt.hash(password1, salt, function (err, hash) {
-      res.locals.password = hash;
-      next();
-    });
-  });
-};
-
-const addUser = (req, res, next) => {
-  const text = `INSERT INTO ubiquitous_spoon.users (username, password, name, email, vegan, vegetarian, gluten_free)
-  VALUES('${req.body.username}', '${res.locals.password}', '${req.body.name}', '${req.body.email}', '${req.body.vegan}', '${req.body.vegetarian}', '${req.body.glutenFree}');`;
-  pool.query(text, (err, response) => {
-    if (err) {
-      return next({ log: 'addUser', message: { err: 'Error in addUser' } });
-    }
-    next();
-  });
-};
-
-const createSession = (req, res, next) => {
-  const ssid = uuidv4();
-  const text = `INSERT INTO ubiquitous_spoon.sessions (id, created_at) VALUES('${ssid}', current_timestamp);`;
-  pool.query(text, (err, response) => {
-    if (err) {
-      return next({ log: 'createSession', message: { err: 'Error in createSession' } });
-    }
-    res.cookie('ssid', ssid, { httpOnly: true });
-    next();
-  });
-};
-
 // Check whether username already exists
-app.post('/api/signup', checkUsername, checkPassword, addUser, createSession, (req, res) => {
-  res.status(200).json({ success: true, username: res.locals.username });
-});
+app.use('/api/signup', usersRouter);
 
 // Error handler
 const errorHandler = (err, req, res, next) => {
